mmu+elf_load

+++ inc/misc.h
@@ -1,4 +1,4 @@
 #ifndef _UC32SIM_MISC_H
 #define _UC32SIM_MISC_H
-#define static_assert(expr) do {switch ((int)expr) {case 0: case (int)expr: break;} }while(0)
+#define static_assert(expr) do {switch (expr) {case 0: case expr: break;} }while(0)
 #endif
